assignment/file-explorer/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
google-auth==2.23.4
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
passlib[bcrypt]==1.7.4
